requirements.txt
python-dotenv
groq
requests
bs4
gradio